+++ NLPCPayRollApp/Models/PayrollComponents.cs
@@ -7,30 +7,30 @@ namespace NLPCPayRollApp.Models
     {
         public int Id { get; set; }
         [DisplayName("Basic Salary")]
-        public int BasicSalary { get; set; }
-        public int Housing { get; set; }
-        public int Transport { get; set; }
-        public int Furniture { get; set; }
-        public int Entertainment { get; set; }
-        public int Lunch { get; set; }
-        public int Passage { get; set; }
-        public int Dressing { get; set; }
-        public int Bonus { get; set; }
+        public double BasicSalary { get; set; }
+        public double Housing { get; set; }
+        public double Transport { get; set; }
+        public double Furniture { get; set; }
+        public double Entertainment { get; set; }
+        public double Lunch { get; set; }
+        public double Passage { get; set; }
+        public double Dressing { get; set; }
+        public double Bonus { get; set; }
         [DisplayName("13th Month")]
-        public int ThirteenthMonth { get; set; }
-        public int Utility { get; set; }
+        public double ThirteenthMonth { get; set; }
+        public double Utility { get; set; }
         [DisplayName("Other Allowances")]
-        public int OtherAllowances { get; set; }
-        public int NHF { get; set; }
-        public int NHIS { get; set; }
+        public double OtherAllowances { get; set; }
+        public double NHF { get; set; }
+        public double NHIS { get; set; }
         [DisplayName("National Pension Scheme")]
-        public int NPS { get; set; }
+        public double NPS { get; set; }
         [DisplayName("Life Assurance")]
-        public int LifeAssurance { get; set; }
+        public double LifeAssurance { get; set; }
         [DisplayName("Gross Income")]
-        public int GrossIncome { get; set; }
+        public double GrossIncome { get; set; }
         [DisplayName("Tax Payable")]
-        public int TaxPayable { get; set; }
+        public double TaxPayable { get; set; }
         public int CadreLevelId { get; set; }
 
     }
